+++ releasenotes.md
@@ -3,6 +3,7 @@
 #### DOOM Retro v4.0.1
 
 * A bug has been fixed whereby sprites would be incorrectly drawn in front of masked midtextures in some instances.
+* The external automap now displays correctly when the `am_external` CVAR is `on`.
 * Minor changes have been made to text that is output to the console.
 * Music will now play again when loading a savegame created using *DOOM Retro v3.6* or *v3.6.1*.
 * If the player starts a new game from the menu while playing *E1M4B: Phobos Mission Control* or *E1M8B: Tech Gone Bad*, the correct map will now be loaded.